Fixed some merge problems
