+++ src/routes/+page.svelte
@@ -1,49 +1,65 @@
 <script>
-    export let data;
-    let persons = data.persons;
-    import Brief from '$lib/briefForm.svelte';
-    import DesktopCard from '../lib/desktopCard.svelte';
-
-    function checkAvatarImage(avatar) {
-        if (avatar.includes('https')) {
-            return true
-        }
-
-        return false
+    import Package from "$lib/package/Package.svelte";
+    // export let data;
+  </script>
+  
+  <h1>FDND</h1>
+  
+  <main>
+    <Package
+      colors={{
+        topLeft: "#f2bd80",
+        topRight: "#f2bd80",
+        left: "#ad8455",
+        right: "#f5bf82",
+        front: "#e3b178",
+        back: "#ad8455",
+        border: "#c9a574",
+      }}
+      size={8}
+      rotation={{ x: "-10", y: "-15" }}
+      title={"Squad C"}
+      href={"/squad/3"}
+    />
+  
+    <Package
+    colors={{
+      topLeft: "#f2bd80",
+      topRight: "#f2bd80",
+      left: "#f5bf82",
+      right: "#ad8455",
+      front: "#e3b178",
+      back: "#ad8455",
+      border: "#c9a574",
+    }}
+    size={8}
+      rotation={{ x: "-10", y: "15" }}
+      title={"Squad D"}
+      href={"/squad/4"}
+    />
+  </main>
+  
+  <style>
+    h1 {
+      width: 100%;
+      text-align: center;
+      font-size: 3rem;
+      text-transform: uppercase;
+      font-family: "Open Sans", "Helvetica Neue", sans-serif;
+      letter-spacing: 2px;
+      font-weight: 800;
+      color: white;
+      margin-top: 2rem;
     }
-
-    console.log({data});
-</script>
-
-<style>
-    @import url('https://fonts.googleapis.com/css2?family=Italianno&display=swap');
-
-    body{
-        background-color: #33293A;
-    }
-
-    header{
-        font-size: 80px;
-        color: white;
-        font-family: "Italianno", sans-serif;
-        margin-left: 0.5em;
-        margin-block: 1em;
-    }
-
+  
     main {
-        line-height: 1.5;
-        font-family: sans-serif;
-        font-size: 16px;
-        background-color: #33293A;
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      gap: 1rem;
+      justify-content: center;
+      align-items: center;
+      padding: 1rem;
+      width: 100%;
+      height: 100%;
     }
-</style>
-
-<body>
-    <header> 
-        squadpage 
-    </header>
-    <main>
-        <DesktopCard {persons}/>
-    </main>
-</body>
-<Brief/>
+  </style>
